Fix ALpine based docker image
ec5a505 switched the base image to "alpine:latest". Yet, the Erlang version
used to build capwap-dp was based on bitwalker/alpine-erlang:20.1 which
differs from the official Alpine packages of Erlang. This caused unresolved
symbols in the created binary of capwap-dp.

This commit uses official Alpine packages of Erlang to build and run
capwap-dp. Also, the explicit use of alpine:3.7 brings capwap-dp in sync with
its capwap-ac counterpart.