Tweak Docker operation
First, this commit upgrades the base image to Alpine:3.8.

Second, `capwap-dp` is built with debug symbols. The debug symbols are
split out of the `capwap-dp` binary, but are kept in the docker image.
This allows to analyse crashes etc without adding any overhead to the
binary.

Third, the docker image is using now a `entrypoint.sh` script to
actually start `capwap-dp` … plus sugar ontop of it. The reason is
simple: `capwap-dp` creates a `tap0` interface which is in state "down".
So, usually other mechanisms like if-up-down scripts from the Linux
distribution of choice is kicking in and setting up `tap0` or other
interfaces (bridges) etc. Inside docker, no such mechanisms exist. One
option is to use `udev`. A simple `entrypoint.sh` will do just fine.

The `entrypoint.sh` will launch the given command in the background,
check for an optional helper script named /etc/capwap-dp/capwap-dp-up.sh
and execute it, then wait for the background process to finish.

And finally, this commit changes the default `capwap-dp.conf` file: Both
the "listen" and the "forward" directive introduce dedicated namespaces.
This makes `entrypoint.sh` more complicated per default.