Improve building Docker image
First of all, this commit removes the built-time dependency to
`shtoolize`: It is not needed anyway - noone needs a good looking build
log.

Secondly, `--enable-debug` is removed from configuration: In addition to
compile `capwap-dp` with "-O2 -g" (optimized code, but debug symbols),
`--enable-debug` also enables debug logging - which is a bit too much
for production. So, this commit also reverts that part of 0399475.