Improve entrypoint.sh for Docker image
One way to use the `capwap-dp` image is as part of a Kubernetes
deployment. As such, the `/etc/capwap-dp/capwap-dp-up.sh` helper script
is provided as "Configmap". A "Configmap" is not executable.

By explicitely using `sh` to process `/etc/capwap-dp/capwap-dp-up.sh`
works also in a Kubernetes environment.

Also, the `-e` flag of the `entrypoint.sh` was changed towards `-x`:
Treating errors as fatal exit triggers in the helper scripts or when
bringin up the tap0 interface after launching `capwap-dp` is a bit too
strict. By using `-x`, we get more visibility.