fix for speculative use-after-free problem
Build on newer userspace-rcu library resulted in spurious segfault.
It is possible that call_rcu with only the current thread holding a
look are executed immediatly (or at least sufficiently fast) that the
resulting free() from the call happens before the lock is released.
This could potentially result in the current worker to access freed
data.

Change the logic to not access the data anymore after invking call_rcu
to release the pointer.