replace cds_hlist_for_each_entry_rcu_2 with cds_hlist_for_each_entry_rcu
The compiler (gcc 10.2.1) decides that the test for the NULL pointer at the
end of the list in cds_hlist_for_each_entry_rcu_2 can never succeed and turn
the into and endless loop. That will of course the loop to overrun the end
and dereference the NULL pointer.

cds_hlist_for_each_entry_rcu does not suffer from that problem.